Reassign project repository responsibilities to project service
