+++ src/project/project.service.spec.ts
@@ -3,12 +3,12 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { PinoLogger } from 'nestjs-pino';
 import { DbException } from '../utils/exceptions/database.exception';
 import { SerializationModule } from '../utils/serialization/serialization.module';
-import { ProjectCustomRepository } from './project.repository';
+import { QueryCreator } from './project.query.creator';
 import { ProjectService } from './project.service';
 
 describe('ProjectService', () => {
   let service: ProjectService;
-  const projectCustomRepositoryMock = { findOne: jest.fn() };
+  const queryCreatorMock = { findOne: jest.fn() };
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
@@ -19,8 +19,8 @@ describe('ProjectService', () => {
           useValue: { debug: jest.fn(), setContext: jest.fn() },
         },
         {
-          provide: ProjectCustomRepository,
-          useValue: projectCustomRepositoryMock,
+          provide: QueryCreator,
+          useValue: queryCreatorMock,
         },
       ],
       imports: [SerializationModule],
@@ -30,7 +30,7 @@ describe('ProjectService', () => {
   });
 
   afterEach(() => {
-    projectCustomRepositoryMock.findOne.mockReset();
+    queryCreatorMock.findOne.mockReset();
   });
 
   it('should be defined', () => {
@@ -41,7 +41,7 @@ describe('ProjectService', () => {
     describe('when no project is found', () => {
       it('should throw a NotFound exception', async () => {
         const noMatchingId = 155;
-        projectCustomRepositoryMock.findOne.mockResolvedValue(undefined);
+        queryCreatorMock.findOne.mockResolvedValue(undefined);
         await service.findOne(noMatchingId).catch((error) => {
           expect(error).toBeInstanceOf(NotFoundException);
           expect(error.response.message).toBe('Not Found');
@@ -52,7 +52,7 @@ describe('ProjectService', () => {
 
     describe('when an exception is thrown by repository method', () => {
       it('should re-trow a db exception', async () => {
-        projectCustomRepositoryMock.findOne.mockRejectedValue(new Error());
+        queryCreatorMock.findOne.mockRejectedValue(new Error());
         const anyId = 155;
         await service.findOne(anyId).catch((error) => {
           expect(error).toBeInstanceOf(DbException);
